requirements.txt
asn1crypto==0.24.0
aws==0.2.5
awscli==1.16.150
awsebcli==3.15.0
bcrypt==3.1.6
blessed==1.15.0
boto==2.49.0
botocore==1.12.131
bottle==0.12.16
cached-property==1.5.1
cachetools==3.1.0
cement==2.8.2
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.7
dnspython==1.16.0
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
fabric==2.4.0
Flask==1.0.3
future==0.16.0
get==2019.4.13
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
gunicorn==19.9.0
httplib2==0.12.1
idna==2.7
invoke==1.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==2.6.0
MarkupSafe==1.1.1
oauthlib==3.0.1
ordereddict==1.1
paramiko==2.4.2
pathspec==0.5.9
post==2019.4.13
prettytable==0.7.2
prompt-toolkit==2.0.9
public==2019.4.13
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pymessenger==0.0.7.0
pymongo==3.7.2
PyNaCl==1.3.0
python-dateutil==2.8.0
python-telegram-bot==12.1.1
PyYAML==3.13
query-string==2019.4.13
redis==3.2.1
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rq==1.0
rsa==4.0
s3transfer==0.2.0
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.4
wit==5.1.0
schedule==0.6.0
telebot==0.0.3
pyTelegramBotAPI==3.6.6